\name{FREddyPro-package}
\alias{FREddyPro-package}
\alias{FREddyPro}
\docType{package}
\title{
  \packageTitle{FREddyPro}
}
\description{
  \packageDescription{FREddyPro}
}
%% \details{ 
%%   The DESCRIPTION file:
%%   \packageDESCRIPTION{FREddyPro}
%%   \packageIndices{FREddyPro}
%% }
\author{
  \packageAuthor{FREddyPro}
  
  Maintainer: \packageMaintainer{FREddyPro}
}
\examples{
  library(FREddyPro)
}
