\name{spectra}
\alias{spectra}
\docType{data}
\title{
  Ensemble spectra from semi-enclosed IRGA at Harwood forest
}
\description{
  Standard EddyPro ensemble spectra for Harwood forest, Northumberland,
  Great Bridtain. The eddy covariance tower is managed by Forest Research
}
\usage{data("spectra")}
\examples{
## Load the data
data(spectra)

## Explore the variables in the data frame
str(spectra)
}
